+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+- js - added diagram based render config options
+  - https://github.com/StateSmith/StateSmith/issues/23
+  
+---
+
 ## [0.8.11-alpha]
 ### Changes
 - Enabled automatic name conflict resolution by default